Fix EZP-24457: Change Request injection to RequestStack
> https://jira.ez.no/browse/EZP-24457

Request injection has been deprecated as of Symfony 2.4 and will be removed in 3.0.
Instead we must use RequestStack, which contains all ongoing requests (including sub-requests).

This PR fixes all `@request` injection.
It also introduces a trait, `RequestStackAware`, to ease access to the current request.
For controllers, signatures have been changed to add the request as an argument to actions,
[as described in Symfony documentation](http://symfony.com/doc/current/book/controller.html#book-controller-request-argument).